+++ Source/Framework/ZyGames.Framework/Data/Sql/SqlDataProvider.cs
@@ -220,7 +220,7 @@ namespace ZyGames.Framework.Data.Sql
         /// <param name="commandText"></param>
         /// <param name="parameters"></param>
         /// <returns></returns>
-        public override int ExecuteNonQuery(int identityId, CommandType commandType, string tableName, string commandText, params IDataParameter[] parameters)
+        public override int ExecuteNonQuery(long identityId, CommandType commandType, string tableName, string commandText, params IDataParameter[] parameters)
         {
             SqlStatement statement = new SqlStatement();
             statement.IdentityID = identityId;
@@ -239,7 +239,7 @@ namespace ZyGames.Framework.Data.Sql
         /// <param name="identityId"></param>
         /// <param name="command"></param>
         /// <returns></returns>
-        public override SqlStatement GenerateSql(int identityId, CommandStruct command)
+        public override SqlStatement GenerateSql(long identityId, CommandStruct command)
         {
             command.Parser();
             SqlStatement statement = new SqlStatement();
